refactor : User Domain refactoring

+++ backend/dduishu/src/main/java/com/b208/dduishu/domain/user/dto/request/UserRankingInfo.java
@@ -12,7 +12,7 @@ public class UserRankingInfo {
 
     private Long userId;
     private String nickname;
-    private float cumulativeDistance;
+    private double cumulativeDistance;
     private int level;
     private boolean isFollower;
